+++ CHANGELOG.md
@@ -5,6 +5,7 @@ omnibus-gitlab repository.
 
 11.3.0
 
+- Support Redis hz parameter
 - Support Redis tcp-backlog parameter
 - Disable SSL compression by default when using gitlab-psql or gitlab-geo-psql
 - Increase Sidekiq RSS memory limit from 1 GB to 2 GB